% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefs.R
\name{fd_coefs}
\alias{fd_coefs}
\title{Calculate finite difference coefficients at artificial stencil points}
\usage{
fd_coefs(order, stencil)
}
\arguments{
\item{order}{order of derivative}

\item{stencil}{locations of sampled points}
}
\value{
a numeric vector of finite difference coefficients corresponding to stencil points.
}
\description{
Calculate finite difference coefficients at artificial stencil points
}
\examples{
coefs <- fd_coefs(-3:3, 2)
}
