% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finite.R
\name{finite_diff}
\alias{finite_diff}
\title{Calculate finite difference of a function}
\usage{
finite_diff(f, x, order, stencil, h = 1)
}
\arguments{
\item{f}{a function or a numeric vector}

\item{x}{position to calculate the finite difference}

\item{order}{order of derivative}

\item{stencil}{sampled points relative to x}

\item{h}{an epsilon}
}
\value{
a numeric value, the derivative of f(x)
}
\description{
Calculate finite difference of a function
}
